+++ jvm/src/main/scala/org/sireum/hamr/codegen/microkit/util/JvmMicrokitUtil.scala
@@ -1,0 +1,14 @@
+// #Sireum
+package org.sireum.hamr.codegen.microkit.util
+
+import org.sireum._
+
+object JvmMicrokitUtil {
+  @pure def getMicrokitVersions: Map[String, String] = {
+    return JvmMicrokitUtilExt.getMicrokitVersions
+  }
+}
+
+@ext object JvmMicrokitUtilExt {
+  @pure def getMicrokitVersions: Map[String, String] = $
+}